requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
sentence-transformers>=2.6
faiss-cpu>=1.7
scikit-learn>=1.3
python-dotenv>=1.0
torch>=2.1